requirements.txt
streamlit
flask-login
flask-wtf
matplotlib
numpy
pandas
pillow
psycopg2-binary
